requirements.txt
robotframework==4.1.3
robotframework-pabot~=2.1.0
robotframework-seleniumlibrary==5.1.3
openpyxl==3.0.9
pydrive~=1.3.1
google-auth-oauthlib
selenium
requests~=2.26.0
bs4~=0.0.1
beautifulsoup4~=4.10.0
httplib2~=0.15.0
robotframework-dependencylibrary
robotframework-tidy
pandas==1.4.4
robotframework-csvlibrary==0.0.5
